db문제 해결 및 글 수정에서 이미지 업로드 기능 구현
