+++ src/main/java/com/mandarin/petching/domain/Member.java
@@ -1,6 +1,8 @@
 package com.mandarin.petching.domain;
 
+import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
+import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import com.mandarin.petching.dto.MemberFormDto;
 import lombok.Getter;
 import lombok.Setter;
@@ -12,6 +14,7 @@ import java.util.List;
 
 @Entity
 @Getter @Setter
+@JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class , property = "id")
 public class Member {
     @Id
     @Column(name = "member_id")
@@ -20,6 +23,7 @@ public class Member {
 
     private String userName;
 
+//    @JsonManagedReference
     @OneToMany(mappedBy = "member", cascade = CascadeType.REMOVE)
     private List<Pet> petList = new ArrayList<>();
 
@@ -41,7 +45,7 @@ public class Member {
     @Enumerated(EnumType.STRING)
     private Role role;
 
-    @JsonManagedReference
+//    @JsonManagedReference
     @OneToMany(mappedBy = "member", cascade = CascadeType.REMOVE)
     private List<Board> board = new ArrayList<>();
