+++ src/main/java/com/mandarin/petching/domain/Reply.java
@@ -1,5 +1,9 @@
 package com.mandarin.petching.domain;
 
+import com.fasterxml.jackson.annotation.JsonBackReference;
+import com.fasterxml.jackson.annotation.JsonIdentityInfo;
+import com.fasterxml.jackson.annotation.JsonManagedReference;
+import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,6 +19,7 @@ import javax.persistence.*;
 @Entity
 @Getter
 @Setter
+@JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class , property = "id")
 public class Reply {
 
     @Id
@@ -22,6 +27,7 @@ public class Reply {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+//    @JsonBackReference
     @ManyToOne
     @JoinColumn(name="board_id", referencedColumnName = "board_id")
     private Board board;
@@ -33,6 +39,7 @@ public class Reply {
     @CreatedDate
     private LocalDateTime reRegDate;
 
+//    @JsonBackReference
     @ManyToOne
     @JoinColumn(name="user_id")
     private Member member;
